move getLogSql to metadataTableUtils.ts
