+++ console/src/metadata/queryUtils.ts
@@ -12,9 +12,8 @@ import { LocalEventTriggerState } from '../components/Services/Events/EventTrigg
 import { LocalScheduledTriggerState } from '../components/Services/Events/CronTriggers/state';
 import { LocalAdhocEventState } from '../components/Services/Events/AdhocEvents/Add/state';
 import { RemoteRelationshipPayload } from '../components/Services/Data/TableRelationships/RemoteRelationships/utils';
-import { Driver, currentDriver, dataSource } from '../dataSources';
+import { Driver, currentDriver } from '../dataSources';
 import { ConsoleState } from '../telemetry/state';
-import { getRunSqlQuery } from '../components/Common/utils/v1QueryUtils';
 
 export const metadataQueryTypes = [
   'add_source',
@@ -628,48 +627,6 @@ export const getSetConsoleStateQuery = (
 export const getConsoleStateQuery = {
   type: 'get_catalog_state',
   args: {},
-};
-
-export const getLogSql = (
-  queryType: 'select' | 'count',
-  triggerName: string,
-  table: QualifiedTable,
-  relationships: string[],
-  limit?: number,
-  offset?: number
-) => {
-  let eventType = 'cron';
-  // FIXME: test and change for scheduled events
-  if (relationships[0].includes('scheduled')) {
-    eventType = 'scheduled';
-  }
-
-  const relTable: QualifiedTable = {
-    schema: 'hdb_catalog',
-    name: `hdb_${eventType}_events`,
-  };
-
-  if (!dataSource.getInvocationLogSql) {
-    return;
-  }
-
-  const sql = dataSource.getInvocationLogSql(
-    'cron',
-    table,
-    relTable,
-    triggerName,
-    limit,
-    offset
-  );
-
-  // todo: wait for API / write new SQL for this.
-  // if (queryType === 'count') {
-  //   sql += ';';
-  // } else {
-  //   sql += ` LIMIT ${limit ?? 10} OFFSET ${offset ?? 0};`;
-  // }
-
-  return getRunSqlQuery(sql, 'default');
 };
 
 export const getInvocationLogsQuery = (
